% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing.R
\name{add_missing}
\alias{add_missing}
\title{add_missing}
\usage{
add_missing(data)
}
\arguments{
\item{data}{dataframe to test and convert}
}
\description{
Used to add missing data to input files and customize format
}
