% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_data.R
\name{aggregate_data}
\alias{aggregate_data}
\title{aggregate_data}
\usage{
aggregate_data(data = NULL, col_agg = "class")
}
\arguments{
\item{data}{dataframe to test and convert}

\item{col_agg}{Default = "class". Column to remove and then aggregate by.}
}
\description{
Used to add missing data to input files and customize format
}
