% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diff.R
\name{calculate_diff}
\alias{calculate_diff}
\title{calculate_diff}
\usage{
calculate_diff(
  data = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  xRef = NULL,
  xDiff = NULL,
  diff_type = "both",
  diff_type_x = "both",
  diff_text_percent = "diffPrcnt",
  diff_text_absolute = "diffAbs",
  diff_text_percent_x = "xdiffPrcnt",
  diff_text_absolute_x = "xdiffAbs"
)
}
\arguments{
\item{data}{Default = NULL.}

\item{scenRef}{Default = NULL. Reference Scenario}

\item{scenDiff}{Default = NULL. Difference Scenarios}

\item{xRef}{Default = NULL. Reference x}

\item{xDiff}{Default = NULL. Difference x}

\item{diff_type}{Default = "both". One of "absolute", "percent", "both".}

\item{diff_type_x}{Default = "both". One of "absolute", "percent", "both".}

\item{diff_text_percent}{Default = "_diffPrcnt"}

\item{diff_text_absolute}{Default = "_diffAbs"}

\item{col}{Default = NULL.}
}
\description{
Calculate difference between scenarios or time periods
}
