% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{chart}
\usage{
chart(
  data = NULL,
  col_agg = "class",
  chart_type = "all",
  aspect_ratio = 0.75,
  size = 1.5,
  size_text = 10,
  theme = NULL,
  ncol = NULL,
  scales = "free_y",
  save = TRUE,
  folder = getwd()
)
}
\arguments{
\item{data}{Default = NULL. Dataframe to test and convert.}

\item{col_agg}{Default = "class". Column to remove and then aggregate by.}

\item{chart_type}{Default = "all". Choices one or more of "all", "lines", "bar", "lines_diff", "bar_diff"}

\item{aspect_ratio}{Default = 0.75. aspect ratio}

\item{size}{Default = 1.5. line size}

\item{size_text}{Default = NULL. Text size}

\item{theme}{Default = NULL.}

\item{ncol}{Default = 3. Number of columns.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{save}{Default = TRUE. Save plots.}

\item{folder}{Default = getwd().}
}
\description{
Used to add missing data to input files and customize format
}
