% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampleMapDataParam}
\alias{exampleMapDataParam}
\title{Example data by param to plot}
\format{
R table or .csv
}
\source{
example data by Param
}
\usage{
exampleMapDataParam
}
\description{
Example data by param to plot
}
\examples{
\dontrun{
 library(rchart);
 data <- rchart::exampleDataParam
}
}
\keyword{datasets}
