% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute.R
\name{plot_absolute}
\alias{plot_absolute}
\title{plot_absolute}
\usage{
plot_absolute(data = NULL, theme = ggplot2::theme_bw())
}
\arguments{
\item{data}{Default = NULL.}

\item{theme}{ggplot theme to use}
}
\description{
generate chart plot for absolute value
}
