% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_difference.R
\name{plot_class_difference}
\alias{plot_class_difference}
\title{plot_class_difference}
\usage{
plot_class_difference(
  data = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  ncol = NULL,
  diff_text = NULL,
  scales = "free",
  diff_type = "bar",
  size = 1.5,
  break_interval = NULL,
  include_points = FALSE,
  summary_line = FALSE,
  data_agg_ref = NULL,
  data_agg_diff = NULL,
  palette = NULL
)
}
\arguments{
\item{data}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{theme}{Default = NULL}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{ncol}{Default = NULL. Numbers of columns in the wrapped plots.}

\item{diff_text}{Default = NULL. Text to remove from diff scenario names.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{diff_type}{Default = "bar". Choose between "bar" or "line"}

\item{size}{Default = 1.5. Line size}

\item{break_interval}{Default = NULL. Intervals between x breaks starting from first x point.}

\item{include_points}{Default = FALSE. Add data points to all line charts.}

\item{summary_line}{Default = FALSE. Add parameter summary line to all bar charts.}

\item{data_agg_ref}{Default = NULL. Aggregated param data for the summary line on the scenRef bar chart.}

\item{data_agg_diff}{Default = NULL. Aggregated param diff data for the summary lines on the scenDiff bar charts.}

\item{palette}{Default = NULL. Named vector with custom palette colors (can include classes, regions, and/or scenarios; class colors will be used if provided)}
}
\description{
generate chart plot for absolute difference and percent difference
}
