% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_waterfall.R
\name{plot_class_waterfall}
\alias{plot_class_waterfall}
\title{plot_class_waterfall}
\usage{
plot_class_waterfall(
  data_diff = NULL,
  data_agg = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  diff_text = NULL,
  scales = "free",
  diff_type = "bar",
  size = 1.5,
  break_interval = NULL,
  include_points = FALSE,
  summary_line = FALSE,
  wf_x = NULL,
  rect_width = 0.7,
  aspect_ratio = 0.6,
  horizontal_lines = TRUE,
  lty = 2,
  fill_colors = NULL,
  totals_fill_color = "gray90",
  palette = NULL,
  vertical_dim = NULL,
  ylim = NULL,
  single_chart = F,
  scen_order = NULL
)
}
\arguments{
\item{data_diff}{Default = NULL.}

\item{data_agg}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{theme}{Default = NULL}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{diff_text}{Default = NULL. Text to remove from diff scenario names.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{diff_type}{Default = "bar". Choose between "bar" or "line"}

\item{size}{Default = 1.5. Line size}

\item{break_interval}{Default = NULL. Intervals between x breaks starting from first x point.}

\item{include_points}{Default = FALSE. Add data points to all line charts.}

\item{summary_line}{Default = FALSE. Add parameter summary line to all bar charts.}

\item{wf_x}{Default = NULL. Year (or x value) for which to make the waterfall plot.}

\item{rect_width}{Default = 0.7. Width of the rectangles in the waterfall plot.}

\item{aspect_ratio}{Default = 0.6. Aspect ratio of the plot. Default is lower (more horizontal) than other rchart default}

\item{horizontal_lines}{Default = TRUE. Whether to include horizontal lines between rectangles}

\item{lty}{Default = 2. Line type for the horizontal lines between rectangles}

\item{fill_colors}{Default = NULL. Vector of colors for rectangles. If not specified, uses jgcricolors corresponding to classes}

\item{totals_fill_color}{Default = "gray90". Color of the param total bars for the ref and diff scenarios}

\item{palette}{Default = NULL. Named vector with custom palette colors (can include classes, regions, and/or scenarios; class colors will be used if provided)}

\item{vertical_dim}{Default = NULL. Column to use for vertical stacking within classes}

\item{ylim}{Default = NULL. Y-axis limits}

\item{single_chart}{Default = F. If there are multiple diff scenarios, include them all in the same chart?}

\item{scen_order}{Default = F. Order of scenarios to chart if single chart option is selected}
}
\description{
generate chart plot for absolute difference and percent difference
}
