% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_difference.R
\name{plot_difference}
\alias{plot_difference}
\title{plot_difference}
\usage{
plot_difference(
  data = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  plot_type = "bar",
  theme = NULL
)
}
\arguments{
\item{data}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{plot_type}{Default = "bar". Choose one of "bar" or "line",}

\item{theme}{Default = NULL}
}
\description{
generate chart plot for absolute difference and percent difference
}
