% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_line_absolute.R
\name{plot_line_absolute}
\alias{plot_line_absolute}
\title{plot_line_absolute}
\usage{
plot_line_absolute(
  data = NULL,
  aspect_ratio = 0.75,
  size = 1.5,
  size_text = 10,
  theme = NULL,
  ncol = 3,
  scales = "free_y"
)
}
\arguments{
\item{data}{Default = NULL.}

\item{aspect_ratio}{Default = 0.75. aspect ratio}

\item{size}{Default = 1.5. Line size}

\item{size_text}{Default = 10. text size}

\item{theme}{Default = NULL.}

\item{ncol}{Default = 3. Number of columns.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}
}
\description{
generate summary plot
}
