% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reg_absolute.R
\name{plot_reg_absolute}
\alias{plot_reg_absolute}
\title{plot_reg_absolute}
\usage{
plot_reg_absolute(
  data = NULL,
  size = 1.5,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  scales = "free_y"
)
}
\arguments{
\item{data}{Default = NULL.}

\item{size}{Default = 1.5. Line size}

\item{theme}{Default = NULL.}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}
}
\description{
generate summary plot
}
