% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rchart.R
\docType{package}
\name{rchart}
\alias{rchart}
\title{rchart}
\description{
\itemize{
\item Github: https://github.com/JGCRI/rchart
\item Webpage: https://jgcri.github.io/rchart/
\item Cheatsheet: https://github.com/JGCRI/rmap/blob/master/cheatsheet.pdf}
}
