% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{chart}
\usage{
chart(
  data = NULL,
  col_agg = "class",
  aspect_ratio = 0.75,
  size = 10,
  size_title = NULL,
  theme = NULL
)
}
\arguments{
\item{data}{Default = NULL. Dataframe to test and convert.}

\item{col_agg}{Default = "class". Column to remove and then aggregate by.}

\item{aspect_ratio}{Default = 0.75. aspect ratio}

\item{size}{Default = 10. text size}

\item{size_title}{Default = NULL. size_title text}

\item{theme}{Default = NULL.}

\item{chart_type}{Default = "all". Choices one or more of "all", "lines", "bar", "lines_diff", "bar_diff"}
}
\description{
Used to add missing data to input files and customize format
}
