% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart.R
\name{chart}
\alias{chart}
\title{chart}
\usage{
chart(
  data = NULL,
  col_agg = "class",
  chart_type = "all",
  aspect_ratio = 0.75,
  size = 0.75,
  size_text = 15,
  theme = NULL,
  ncol = NULL,
  scales = "free_y",
  save = TRUE,
  show = FALSE,
  title = TRUE,
  folder = getwd(),
  scenRef = NULL,
  scenDiff = NULL,
  xRef = NULL,
  xDiff = NULL,
  diff_type = "both",
  diff_type_x = "both",
  diff_text_percent = "diffPrcnt",
  diff_text_absolute = "diffAbs",
  diff_text_percent_x = "xdiffPrcnt",
  diff_text_absolute_x = "xdiffAbs",
  width = NULL,
  height = NULL,
  append = "",
  break_interval = NULL,
  include_points = FALSE,
  summary_line = FALSE,
  waterfall_x = NULL,
  palette = NULL
)
}
\arguments{
\item{data}{Default = NULL. Dataframe to test and convert.}

\item{col_agg}{Default = "class". Column to remove and then aggregate by.}

\item{chart_type}{Default = "all". Choices one or more of "all", "param_absolute", "class_absolute",
"param_diff_absolute", "class_diff_absolute", "param_diff_percent", "class_diff_percent", "region_absolute"}

\item{aspect_ratio}{Default = 0.75. aspect ratio}

\item{size}{Default = 1.5. line size}

\item{size_text}{Default = 15. Text size}

\item{theme}{Default = NULL.}

\item{ncol}{Default = 3. Number of columns.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{save}{Default = TRUE. Save plots.}

\item{show}{Default = FALSE. Whether to show figure outputs in console.}

\item{title}{Default = TRUE. Used for region or subRegions or to add title to all plots.}

\item{folder}{Default = getwd().}

\item{scenRef}{Default = NULL. Reference Scenario}

\item{scenDiff}{Default = NULL. Difference Scenarios}

\item{xRef}{Default = NULL. Reference x}

\item{xDiff}{Default = NULL. Difference x}

\item{diff_type}{Default = "both". One of "absolute", "percent", "both".}

\item{diff_type_x}{Default = "both". One of "absolute", "percent", "both".}

\item{diff_text_percent}{Default = "_diffPrcnt"}

\item{diff_text_absolute}{Default = "_diffAbs"}

\item{diff_text_percent_x}{Default = "_xdiffPrcnt"}

\item{diff_text_absolute_x}{Default = "_xdiffAbs"}

\item{width}{Default = NULL}

\item{height}{Default = NULL}

\item{append}{Default = ""}

\item{break_interval}{Default = NULL. Intervals between x breaks starting from first x point.}

\item{include_points}{Default = FALSE. Add data points to all line charts.}

\item{summary_line}{Default = FALSE. Add parameter summary line to all bar charts.}

\item{waterfall_x}{Default = NULL. Year (or x value) for which to make waterfall plot. If NULL, latest year will be used}

\item{palette}{Default = NULL. Named vector with custom palette colors (can include classes, regions, and/or scenarios)}
}
\description{
Used to add missing data to input files and customize format
}
