% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_absolute.R
\name{plot_class_absolute}
\alias{plot_class_absolute}
\title{plot_class_absolute}
\usage{
plot_class_absolute(
  data = NULL,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  ncol = 3,
  scales = "free_y",
  size_text = 15,
  break_interval = NULL
)
}
\arguments{
\item{data}{Default = NULL.}

\item{theme}{Default = NULL.}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{ncol}{Default = 3. Number of columns.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{size_text}{Default = 15. Text size}

\item{break_interval}{Default = NULL. Intervals between x breaks starting from first x point.}
}
\description{
generate summary plot
}
