% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_difference.R
\name{plot_class_difference}
\alias{plot_class_difference}
\title{plot_class_difference}
\usage{
plot_class_difference(
  data = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  plot_type = "bar",
  theme = NULL,
  theme_default = ggplot2::theme_bw()
)
}
\arguments{
\item{data}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{plot_type}{Default = "bar". Choose one of "bar" or "line",}

\item{theme}{Default = NULL}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}
}
\description{
generate chart plot for absolute difference and percent difference
}
