% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_class_waterfall.R
\name{plot_class_waterfall}
\alias{plot_class_waterfall}
\title{plot_class_waterfall}
\usage{
plot_class_waterfall(
  data_diff = NULL,
  data_agg = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  diff_text = NULL,
  scales = "free",
  diff_type = "bar",
  size = 1.5,
  break_interval = NULL,
  include_points = FALSE,
  summary_line = FALSE,
  wf_x = NULL,
  rect_width = 0.7,
  aspect_ratio = 0.6,
  horizontal_lines = TRUE,
  lty = 2,
  fill_colors = NULL,
  totals_fill_color = "gray90"
)
}
\arguments{
\item{data_diff}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{theme}{Default = NULL}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{diff_text}{Default = NULL. Text to remove from diff scenario names.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{diff_type}{Default = "bar". Choose between "bar" or "line"}

\item{size}{Default = 1.5. Line size}

\item{break_interval}{Default = NULL. Intervals between x breaks starting from first x point.}

\item{include_points}{Default = FALSE. Add data points to all line charts.}

\item{summary_line}{Default = FALSE. Add parameter summary line to all bar charts.}

\item{wf_x}{Default = NULL. Year (or x value) for which to make the waterfall plot.}

\item{rect_width}{Default = 0.7. Width of the rectangles in the waterfall plot.}

\item{horizontal_lines}{Default = TRUE. Whether to include horizontal lines between rectangles}

\item{lty}{Default = 2. Line type for the horizontal lines between rectangles}

\item{fill_colors}{Default = NULL. Vector of colors for rectangles. If not specified, uses jgcricolors corresponding to classes}

\item{aspec_ratio}{Default = 0.6. Aspect ratio of the plot. Default is lower (more horizontal) than other rchart default}

\item{data_agg_ref}{Default = NULL. Aggregated param data for the summary line on the scenRef bar chart.}

\item{data_agg_diff}{Default = NULL. Aggregated param diff data for the summary lines on the scenDiff bar charts.}
}
\description{
generate chart plot for absolute difference and percent difference
}
