% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diff.R
\name{plot_diff}
\alias{plot_diff}
\title{plot_diff}
\usage{
plot_diff(
  data = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  diff_type = "absolute",
  theme = NULL
)
}
\arguments{
\item{data}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{diff_type}{Default = "absolute". Choose one of "absolute" or "percent"}

\item{theme}{Default = NULL}
}
\description{
generate chart plot for absolute difference and percent difference
}
