% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param_absolute.R
\name{plot_param_absolute}
\alias{plot_param_absolute}
\title{plot_param_absolute}
\usage{
plot_param_absolute(
  data = NULL,
  size = 1.5,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  ncol = 3,
  scales = "free_y",
  break_interval = NULL,
  include_points = FALSE,
  palette = NULL,
  interaction_col_lty = NULL,
  interaction_col_color = NULL
)
}
\arguments{
\item{data}{Default = NULL.}

\item{size}{Default = 1.5. Line size}

\item{theme}{Default = NULL.}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{ncol}{Default = 3. Number of columns.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{break_interval}{Default = NULL. Intervals between x breaks starting from first x point.}

\item{include_points}{Default = FALSE. Add data points to all line charts.}

\item{palette}{Default = NULL. Named vector with custom palette colors (can include classes, regions, and/or scenarios; scenario colors will be used if provided)}

\item{interaction_col_lty}{Default = NULL. Column to use for interaction plot linetype.}

\item{interaction_col_color}{Default = NULL. Column to use for interaction plot color.}
}
\description{
generate summary plot
}
