% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param_difference.R
\name{plot_param_difference}
\alias{plot_param_difference}
\title{plot_param_difference}
\usage{
plot_param_difference(
  data = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  size = 1.5,
  ncol = 3,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  facet_label_diff = "Difference",
  diff_text = NULL,
  scales = "free",
  break_interval = NULL,
  include_points = FALSE,
  palette = NULL
)
}
\arguments{
\item{data}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{size}{Default = 1.5. Line size}

\item{ncol}{Default = 3. Number of columns.}

\item{theme}{Default = NULL}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{facet_label_diff}{Default = Difference}

\item{diff_text}{Default = NULL. Text to remove from diff scenario names.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}

\item{break_interval}{Default = NULL. Intervals between x breaks starting from first x point.}

\item{include_points}{Default = FALSE. Add data points to all line charts.}

\item{palette}{Default = NULL. Named vector with custom palette colors (can include classes, regions, and/or scenarios; scenario colors will be used if provided)}
}
\description{
generate chart plot for absolute difference and percent difference
}
