% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_param_difference.R
\name{plot_param_difference}
\alias{plot_param_difference}
\title{plot_param_difference}
\usage{
plot_param_difference(
  data = NULL,
  scenRef = NULL,
  scenDiff = NULL,
  size = 1.5,
  theme = NULL,
  theme_default = ggplot2::theme_bw(),
  y_label_diff = NULL,
  facet_label_diff = "Difference",
  diff_text = NULL
)
}
\arguments{
\item{data}{Default = NULL.}

\item{scenRef}{Default = NULL.}

\item{scenDiff}{Default = NULL.}

\item{size}{Default = 1.5. Line size}

\item{theme}{Default = NULL}

\item{theme_default}{Default = ggplot2::theme_bw(). Default rchart themes.}

\item{y_label_diff}{Default = NULL}

\item{facet_label_diff}{Default = Difference}

\item{diff_text}{Default = NULL. Text to remove from diff scenario names.}
}
\description{
generate chart plot for absolute difference and percent difference
}
