% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{plot_summary}
\usage{
plot_summary(
  data = NULL,
  aspect_ratio = 0.75,
  size = 10,
  size_title = NULL,
  theme = NULL,
  ncol = 3,
  scales = "free"
)
}
\arguments{
\item{data}{Default = NULL.}

\item{aspect_ratio}{Default = 0.75. aspect ratio}

\item{size}{Default = 10. text size}

\item{size_title}{Default = NULL. size_title text}

\item{theme}{Default = NULL.}

\item{ncol}{Default = 3. Number of columns.}

\item{scales}{Default = "free". Choose between "free", "free_y", "free_x", "fixed"}
}
\description{
generate summary plot
}
