% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{plot_summary}
\usage{
plot_summary(
  data = NULL,
  aspect_ratio = 0.75,
  size = 10,
  size_title = NULL,
  theme = ggplot2::theme_bw()
)
}
\arguments{
\item{data}{data()}

\item{aspect_ratio}{aspect ratio}

\item{size}{text size}

\item{size_title}{size_title text}

\item{theme}{ggplot theme to use}
}
\description{
generate summary plot
}
